# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bissi / Committee / Chit Fund Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/passbook:
    get:
      operationId: getCustomerPassbook
      tags: [customers]
      summary: Get customer digital passbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer passbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passbook"
  /customers/{id}/timeline:
    get:
      operationId: getCustomerTimeline
      tags: [customers]
      summary: Get customer activity timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEntry"
  /customers/{id}/history:
    get:
      operationId: getCustomerHistory
      tags: [customers]
      summary: Get customer full history with all related records
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer full history
          content:
            application/json:
              schema:
                type: object

  # Collectors
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, referenceNumber, name, mobile, status, branchId, createdAt]
      properties:
        id:
          type: integer
        referenceNumber:
          type: string
        name:
          type: string
        mobile:
          type: string
        alternateMobile:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        aadhaar:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        nomineeName:
          type: ["string", "null"]
        nomineeRelation:
          type: ["string", "null"]
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blocked]
        photoUrl:
          type: ["string", "null"]
        referenceName:
          type: ["string", "null"]
        recoveryNotes:
          type: ["string", "null"]
        documents:
          type: ["string", "null"]
        totalTokens:
          type: integer
        totalLoans:
          type: integer
        totalPaid:
          type: number
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, mobile, branchId]
      properties:
        name:
          type: string
        mobile:
          type: string
        alternateMobile:
          type: string
        email:
          type: string
        aadhaar:
          type: string
        pan:
          type: string
        address:
          type: string
        city:
          type: string
        nomineeName:
          type: string
        nomineeRelation:
          type: string
        branchId:
          type: integer
        status:
          type: string
          enum: [active, inactive, blocked]
        photoUrl:
          type: string
        referenceName:
          type: string
        recoveryNotes:
          type: string
        documents:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        alternateMobile:
          type: string
        email:
          type: string
        aadhaar:
          type: string
        pan:
          type: string
        address:
          type: string
        city:
          type: string
        nomineeName:
          type: string
        nomineeRelation:
          type: string
        branchId:
          type: integer
        status:
          type: string
          enum: [active, inactive, blocked]
        photoUrl:
          type: string
        referenceName:
          type: string
        recoveryNotes:
          type: string
        documents:
          type: string
    CustomerListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Passbook:
      type: object
      required: [customer, entries, totalPaid, totalDue]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/PassbookEntry"
        totalPaid:
          type: number
        totalDue:
          type: number
    PassbookEntry:
      type: object
      required: [id, type, description, amount, date]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [payment, loan_disbursed, loan_emi, lottery_win, token_allocation]
        description:
          type: string
        amount:
          type: number
        balance:
          type: ["number", "null"]
        date:
          type: string
    TimelineEntry:
      type: object
      required: [id, type, title, description, date]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        amount:
          type: ["number", "null"]
        date:
          type: string
...